break up messages to scheduler into two different classes
